pyproject.toml
[tool.poetry]
name = "aharas-de2-4-1"
version = "0.1.0"
description = ""
authors = ["hvsio <alicjaharaszczuk@gmail.com>"]
readme = "README.md"
packages = [{include = "aharas_de2"}]

[tool.poetry.dependencies]
python = "<3.13,>=3.8"
sktime = "^0.25.0"
sqlalchemy = "^2.0.25"
sqlalchemy-utils = "^0.41.1"
python-dotenv = "^1.0.0"
psycopg2 = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
